- migrate TimeLine to ts and upgrade d3
