+++ front/package.json
@@ -52,7 +52,7 @@
     "chroma-js": "^2.4.2",
     "classnames": "^2.3.2",
     "country-flag-icons": "^1.5.7",
-    "d3": "^5.16.0",
+    "d3": "^7.8.5",
     "d3-brush": "^3.0.0",
     "d3-drag": "^3.0.0",
     "d3-selection": "^3.0.0",
